fixed efficiency issue
The script was previously turning instances on and off per volume instead of per instance. This was time inefficient for running instances with multiple volumes. Moving functions calls for get_instance_state() and start_instance() outside with for loop solved the issue. The script was tested and runs faster now.